Fixed a bunch of stuff in the azimuth motor control code.  Fixed a major bug in pointing where cos_el and sin_el were exchanged.

+++ mcp/command_struct.h
@@ -53,7 +53,7 @@ struct GainStruct {
 struct VelGainStruct { 
   unsigned short int PE; // prop to velocity error
   unsigned short int PV; // prop to RW velocity
-  unsigned short int SP; // RW velocity Set Point 
+  double SP; // RW velocity Set Point 
 };
 #define LS_OPEN        0x0001
 #define LS_CLOSED      0x0002
